+++ pkgs/standards/tigrbl_auth/tests/i9n/test_rfc8693_token_exchange_endpoint.py
@@ -9,7 +9,7 @@ import pytest
 from fastapi import status
 from httpx import AsyncClient
 
-from tigrbl_auth import encode_jwt
+from tigrbl_auth.rfc.rfc7519 import encode_jwt
 from tigrbl_auth.rfc.rfc8693 import TOKEN_EXCHANGE_GRANT_TYPE, TokenType
