+++ pkgs/standards/tigrbl_auth/tests/unit/test_rfc7523_jwt_profile.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 
 import pytest
 
-from tigrbl_auth import encode_jwt
+from tigrbl_auth.rfc.rfc7519 import encode_jwt
 from tigrbl_auth.errors import InvalidTokenError
 from tigrbl_auth.rfc.rfc7523 import (
     RFC7523_SPEC_URL,
