+++ pkgs/standards/tigrbl_auth/tests/unit/test_rfc7592_client_registration_management.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 
 import pytest
 
-from tigrbl_auth import rfc7591, rfc7592
+from tigrbl_auth.rfc import rfc7591, rfc7592
 
 
 def test_update_and_delete_client_when_enabled() -> None:
