+++ pkgs/standards/tigrbl_auth/tests/unit/test_rfc8932_dns_privacy.py
@@ -12,7 +12,7 @@ validate the stated requirements.
 
 import pytest
 
-from tigrbl_auth import RFC8932_SPEC_URL, enforce_encrypted_dns
+from tigrbl_auth.rfc.rfc8932 import RFC8932_SPEC_URL, enforce_encrypted_dns
 from tigrbl_auth.runtime_cfg import settings
