+++ pkgs/standards/tigrbl_auth/tigrbl_auth/routers/authz/__init__.py
@@ -2,4 +2,5 @@ from fastapi import APIRouter
 
 router = APIRouter()
 
-from . import rfc6749, rfc7662, oidc  # noqa: E402,F401
+from tigrbl_auth.rfc import rfc6749, rfc7662  # noqa: E402,F401
+from . import oidc  # noqa: E402,F401
