+++ pkgs/standards/tigrbl_auth/tigrbl_auth/routers/authz/rfc7662.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException, Request, status
 from ...runtime_cfg import settings
 from ..schemas import IntrospectOut
 from ..shared import _require_tls
-from ...rfc7662 import introspect_token
+from ...rfc.rfc7662 import introspect_token
 
 from . import router
